[BUG] index was not reset, wrong size for integers

+++ include/pattern.h
@@ -13,7 +13,7 @@ typedef struct {
 
 delta *compute_delta_matrix(char *pattern, uint8_t lenght);
 void print_indexes(char *pattern, char *text, uint8_t pattern_lenght, uint32_t text_lenght);
-uint16_t search_pattern(char *pattern, char *text, uint8_t pattern_lenght, uint32_t text_lenght, int32_t *index);
+int32_t search_pattern(char *pattern, char *text, uint8_t pattern_lenght, uint32_t text_lenght, int32_t *index);
 void print_matrix(delta *matrix, uint8_t lenght);
 
 #endif
